requirements.txt
Flask==3.0.0
Pillow==10.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25
alembic==1.13.1
pytest==7.4.4
pytest-cov==4.1.0
